aula aprenda go lang - estrutura fila rabbitmq
