+++ docs/message.json
@@ -1,0 +1,5 @@
+{
+    "filename": "string",
+    "path": "string",
+    "id": "int"
+}
