+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/wagnersilvafilho/aprendagolang/imersao
 
 go 1.22.5
+
+require github.com/rabbitmq/amqp091-go v1.10.0 // indirect
